Implemented summarizer
